backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
pdfplumber==0.11.4
chromadb==0.6.3
google-generativeai==0.8.4
requests==2.32.3
python-dotenv==1.0.1
httpx==0.28.1
